完成时间窗交互从子路径生成，到将子路径转换为四边形框，再将四边形框输出的逻辑，还剩余检测两个四边形框是否重叠的逻辑，以及存储检测结果的逻辑

+++ TimeWindow/timewindow/src/main/java/org/apache/storm/timewindow/poseXYH.java
@@ -2,26 +2,26 @@ package org.apache.storm.timewindow;
 
 public class poseXYH {
 
-	private double x;
-	private double y;
-	private double h;
+	private int x; // cm
+	private int y; // cm
+	private int h;
 	
-	public double getX() {
+	public int getX() {
 		return x;
 	}
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
-	public double getY() {
+	public int getY() {
 		return y;
 	}
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
-	public double getH() {
+	public int getH() {
 		return h;
 	}
-	public void setH(double h) {
+	public void setH(int h) {
 		this.h = h;
 	}
 }
